requirements.txt
boto3==1.40.59
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.123.0
packaging==25.0
pymupdf
